requirements.txt
click==8.0.3
Flask==2.0.2
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mysql-connector-python==8.0.26
pickleshare==0.7.5
python-dateutil==2.8.2
requests==2.26.0
requests-unixsocket==0.2.0
urllib3==1.26.6
webencodings==0.5.1
websocket-client==1.2.1
Werkzeug==2.0.2
numpy
pandas
Pillow
xgboost==1.3.3
scikit-learn
pickle-mixin
flask-moment
Path